# Use to_json when rendering json response