--- a/spec/inertia/rails_mimic_spec.rb
+++ b/spec/inertia/rails_mimic_spec.rb
@@ -6,7 +6,7 @@ RSpec.describe 'rendering when mimicking rails behavior', type: :request, inerti
     it 'has the props' do
       get instance_props_test_path
 
-      expect_inertia.to have_exact_props({'name' => 'Brandon', 'sport' => 'hockey'})
+      expect_inertia.to have_exact_props({name: 'Brandon', sport: 'hockey'})
     end
   end
 
@@ -31,7 +31,7 @@ RSpec.describe 'rendering when mimicking rails behavior', type: :request, inerti
       get default_render_test_path
 
       expect_inertia.to render_component('inertia_rails_mimic/default_render_test')
-      expect_inertia.to include_props({'name' => 'Brian'})
+      expect_inertia.to include_props({name: 'Brian'})
     end
 
     context 'a rendering transformation is provided' do