@BrandonShar 2025-01-07
This is awesome, thanks @alexspeller ! 

Out of curiosity did you do any profiling or metrics for this? It would be cool to see the difference. No need to create something if you didn't.