@BrandonShar 2025-01-08
Wow, that's a big difference! I'm source diving a bit in Rails to understand this better and it looks like this mimics the logic [here](https://github.com/rails/rails/blob/eb6183ffe1ffbfe31573d024ce82ea2a71f5bca0/actionpack/lib/action_controller/metal/renderers.rb#L156)

Am I looking in the wrong place or is there something happening elsewhere first?