@BrandonShar 2025-01-08
@bknoles Only because I'm genuinely not clear on _why_ it's working based on that link I posted above. It looks like the same logic is happening just in different places so I'm hoping to understand a bit more about what's happening here (or be told I'm looking in the wrong place in rails source, it would not be the first time 😄)