@skryukov 2025-01-09
Ok, so the problem lies in Oj's implementation and Rails internally calling `to_json({})`. See https://github.com/ohler55/oj/issues/913

Here's a quick benchmark:

```
require 'active_support/all'
require 'oj'

require 'benchmark/ips'

Oj.optimize_rails

DATA = (1..100_000).map {
  {
    a: rand,
    b: rand,
    c: rand,
    d: rand,
    h: {
      a: rand,
      b: rand,
      h: {
        a: rand,
        b: rand,
      },
    },
  }
}

Benchmark.ips do |x|
  x.report('to_json') { DATA.to_json }
  x.report('to_json({})') { DATA.to_json({}) }
  x.compare!
end
```

```
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +PRISM [arm64-darwin24]
Warming up --------------------------------------
             to_json     1.000 i/100ms
         to_json({})     1.000 i/100ms
Calculating -------------------------------------
             to_json      4.454 (± 0.0%) i/s  (224.49 ms/i) -     23.000 in   5.164198s
         to_json({})      2.097 (± 0.0%) i/s  (476.94 ms/i) -     11.000 in   5.256701s

Comparison:
             to_json:        4.5 i/s
         to_json({}):        2.1 i/s - 2.12x  slower
```

To be honest, I’m struggling with the idea of merging this. On one hand, it will speed things up for Oj users, but on the other hand, Inertia Rails has nothing to do with it.