@skryukov 2025-01-09
On the other hand, the patch isn't obscure or anything, and patching our `Renderer` is not as easy. So, yeah, let's merge it.  
@BrandonShar @bknoles