@alexspeller 2025-01-11
Hi, just coming back to this and that's a really good find, after looking at your link to the rails source I couldn't figure it out either and planned to spend some time digging today - thanks for figuring it out first 😄 

Note that this still offers a good speedup regardless of `Oj` being present or not, when there are unknown keys that are present in the options provided to `to_json` like the rails render pipeline does:

## Without Oj

```rb
require 'active_support/all'
require 'benchmark/ips'

DATA = (1..100_000).map {
  {
    a: rand,
    b: rand,
    c: rand,
    d: rand,
    h: {
      a: rand,
      b: rand,
      h: {
        a: rand,
        b: rand,
      },
    },
  }
}

if defined?(Oj)
  puts "Oj is loaded"
else
  puts "Oj is not loaded"
end

Benchmark.ips do |x|
  x.report('to_json') { DATA.to_json }
  x.report('to_json({})') { DATA.to_json({}) }
  x.report('to_json({prefixes: 123})') { DATA.to_json({prefixes: 123}) }
  x.compare!
end
```

```
Oj is not loaded
ruby 3.3.5 (2024-09-03 revision ef084cc8f4) [arm64-darwin24]
Warming up --------------------------------------
             to_json     1.000 i/100ms
         to_json({})     1.000 i/100ms
to_json({prefixes: 123})
                         1.000 i/100ms
Calculating -------------------------------------
             to_json      1.112 (± 0.0%) i/s  (899.26 ms/i) -      6.000 in   5.417173s
         to_json({})      1.068 (± 0.0%) i/s  (936.55 ms/i) -      6.000 in   5.629989s
to_json({prefixes: 123})
                          0.798 (± 0.0%) i/s     (1.25 s/i) -      4.000 in   5.020716s

Comparison:
             to_json:        1.1 i/s
         to_json({}):        1.1 i/s - 1.04x  slower
to_json({prefixes: 123}):        0.8 i/s - 1.39x  slower
```

## With Oj but without calling Oj.optimize_rails

(exact same benchmark script but with `require 'oj'` at the top)

(almost exact same results as above)

## With Oj but and calling Oj.optimize_rails

(exact same benchmark script but with `require 'oj'` and `Oj.optimise_rails` at the top)


```
Oj is loaded
ruby 3.3.5 (2024-09-03 revision ef084cc8f4) [arm64-darwin24]
Warming up --------------------------------------
             to_json     1.000 i/100ms
         to_json({})     1.000 i/100ms
to_json({prefixes: 123})
                         1.000 i/100ms
Calculating -------------------------------------
             to_json      4.796 (± 0.0%) i/s  (208.53 ms/i) -     24.000 in   5.005977s
         to_json({})      2.041 (± 0.0%) i/s  (490.02 ms/i) -     11.000 in   5.392651s
to_json({prefixes: 123})
                          2.017 (± 0.0%) i/s  (495.70 ms/i) -     11.000 in   5.454995s

Comparison:
             to_json:        4.8 i/s
         to_json({}):        2.0 i/s - 2.35x  slower
to_json({prefixes: 123}):        2.0 i/s - 2.38x  slower
```

It's going to strongly depend on the amount of data and shape of the data, but I think that this should speed up the json render for all users not just Oj users, although the speedup with Oj is more dramatic