@skryukov 2025-01-11
Hey, @alexspeller thanks for digging into this!

> Note that this still offers a good speedup regardless of Oj being present or not

Not quite. The difference in your benchmark comes from the fact that `DATA.to_json({})` creates a new hash every iteration. Rails does that in any case, so a more proper benchmark would be something like this:

```ruby
Benchmark.ips do |x|
  x.report('to_json') { {}; DATA.to_json }
  x.report('to_json({})') { DATA.to_json({}) }
  x.compare!
end
```

which ends up with similar results:

```
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +PRISM [arm64-darwin24]
Warming up --------------------------------------
             to_json     1.000 i/100ms
         to_json({})     1.000 i/100ms
Calculating -------------------------------------
             to_json      1.112 (± 0.0%) i/s  (899.27 ms/i) -      6.000 in   5.396313s
         to_json({})      1.115 (± 0.0%) i/s  (896.66 ms/i) -      6.000 in   5.381196s

Comparison:
         to_json({}):        1.1 i/s
             to_json:        1.1 i/s - 1.00x  slower
```

I still think there are a lot of Oj users out there, so the PR is good to go. Thanks again @alexspeller!