@alexspeller 2025-01-12
The empty hash case is the same as calling without a hash - try this one:

```rb
Benchmark.ips do |x|
  x.report("to_json") { { prefixes: 123 }; DATA.to_json }
  x.report("to_json({prefixes: 123})") { DATA.to_json({ prefixes: 123 }) }
  x.compare!
end
```

```
Oj is not loaded
ruby 3.3.5 (2024-09-03 revision ef084cc8f4) [arm64-darwin24]
Warming up --------------------------------------
             to_json     1.000 i/100ms
to_json({prefixes: 123})
                         1.000 i/100ms
Calculating -------------------------------------
             to_json      0.989 (± 0.0%) i/s     (1.01 s/i) -      5.000 in   5.060382s
to_json({prefixes: 123})
                          0.793 (± 0.0%) i/s     (1.26 s/i) -      4.000 in   5.045480s

Comparison:
             to_json:        1.0 i/s
to_json({prefixes: 123}):        0.8 i/s - 1.25x  slower
```