@skryukov 2025-01-12
> The empty hash case is the same as calling without a hash - try this one:

Inertia only sends `status` and `content_type` options which are removed by Rails, so the only case we should evaluate against is an empty hash 😄